fix: reduce warnings from 424 to ~100 + avx-512 exp/log bugfix
structural changes that eliminate ~320 warnings:
- ax_tensor_t.offset changed from size_t to int64_t (eliminates
  sign-conversion warnings wherever offset is used with int64_t
  loop variables — the bulk of the 424 warnings)
- ax_arena_t given a struct tag (struct ax_arena) so the forward
  declaration in tensor.h matches. fixes all incompatible-pointer
  warnings from arena function calls.

avx-512 fixes:
- exp: use explicit imm=0x08 for vrndscaleps (suppress exceptions)
- log: replace fmadd(r, s, 0) with mul(r, s) — cleaner + avoids
  potential fma-with-zero denormal handling differences

remaining ~100 warnings are -Wsign-conversion in array indexing
(int64_t used as size_t). these need per-site (size_t) casts.

+++ include/axiom/tensor.h
@@ -39,7 +39,7 @@ typedef struct ax_tensor {
     int64_t strides[AX_MAX_DIMS];      /* element stride per dimension */
     int ndim;                          /* number of dimensions */
     ax_dtype_t dtype;                  /* element type */
-    size_t offset;                     /* element offset into storage */
+    int64_t offset;                    /* element offset into storage */
 
     /* autograd fields — used in phase 2, reserved now */
     bool requires_grad;
